Move createStandardFrontstage under FrontstageUtilities
